# apps/backend/src/routes/cc-download.ts
/**
 * @file GET `/api/v1/cc/download/:jamendoId` — same-origin download proxy for a
 * Creative-Commons track.
 *
 * Jamendo's download URL is cross-origin, so a browser ignores the `download`
 * attribute on a link pointing at it and instead navigates to Jamendo's download
 * page — which it saves as a `.html` file, with no control over the name. This
 * proxy re-serves the track's audio from our own origin as an attachment with a
 * `Content-Disposition` filename (`Artist_Album_NN_Title.ext`), so the browser
 * saves a correctly named audio file. It reuses the permanent Jamendo stream URL
 * (the same source the audio player streams) rather than the download page.
 */

import { Readable } from "node:stream";
import type { ReadableStream as NodeWebReadableStream } from "node:stream/web";
import {
  buildCcDownloadFilename,
  JAMENDO_FORMAT_META,
  parseJamendoAudioFormat,
  ROUTE_TEMPLATES,
  swapStreamFormat,
} from "@musiccloud/shared";
import type { FastifyInstance } from "fastify";
import { publicErrorResponse } from "../docs/public-response-schema.js";
import { log } from "../lib/infra/logger.js";
import { sendRateLimitError } from "../lib/infra/rate-limit-response.js";
import { apiRateLimiter, isInternalRequest } from "../lib/infra/rate-limiter.js";
import { getCcTrack } from "../services/cc/jamendo/client.js";

/**
 * Builds the `Content-Disposition` header value for a download: a plain ASCII
 * fallback `filename=` (non-ASCII replaced) plus the RFC 5987 `filename*=` UTF-8
 * form, so names with umlauts/accents survive in modern browsers and degrade
 * gracefully in old ones.
 *
 * @param filename - The desired (possibly non-ASCII) filename.
 * @returns The full `Content-Disposition` header value.
 */
function contentDisposition(filename: string): string {
  const asciiFallback = filename.replace(/[^\x20-\x7e]/g, "_").replace(/"/g, "'");
  return `attachment; filename="${asciiFallback}"; filename*=UTF-8''${encodeURIComponent(filename)}`;
}

/**
 * Registers the CC-download proxy route. Resolves the track by Jamendo id, then
 * streams its audio back as a named attachment in the requested format.
 *
 * @param app - The Fastify instance to register on.
 */
export default async function ccDownloadRoutes(app: FastifyInstance) {
  app.get<{ Params: { jamendoId: string }; Querystring: { format?: string } }>(
    ROUTE_TEMPLATES.v1.ccDownload,
    {
      schema: {
        tags: ["CC"],
        summary: "Download a CC track as a correctly named audio file",
        description:
          "Returns a complete Creative Commons track as a downloadable attachment. `Content-Disposition` provides an ASCII fallback and a UTF-8 filename following `Artist_Album_NN_Title.ext`; unavailable album and track-position segments are omitted. `format=mp31` selects MP3 at `96 kbps`, `format=mp32` selects MP3 at approximately `256 kbps`, `format=ogg` selects Ogg Vorbis, and `format=flac` selects lossless FLAC. The default is `mp32`. Check `track.downloadAllowed` before offering this action; the endpoint returns `403` when downloads are not permitted.",
        params: {
          type: "object",
          required: ["jamendoId"],
          properties: {
            jamendoId: {
              type: "string",
              minLength: 1,
              maxLength: 32,
              pattern: "^[0-9]+$",
              description:
                "Numeric Jamendo track ID. Read `track.jamendoId`, `album.tracks[].jamendoId`, or `artist.topTracks[].jamendoId` from a successful `POST /api/v1/cc/resolve` or `GET /api/v1/share/{shortId}` response.",
            },
          },
          additionalProperties: false,
        },
        querystring: {
          type: "object",
          properties: {
            format: {
              type: "string",
              default: "mp32",
              description:
                "Jamendo delivery format: `mp31` (MP3 96 kbps), `mp32` (MP3 about 256 kbps), `ogg`, or `flac`. Invalid values also fall back to `mp32`.\n\n**Default**: `mp32`.",
            },
          },
          additionalProperties: false,
        },
        response: {
          200: {
            description: "Complete Creative Commons audio file served as an attachment.",
            headers: {
              "Content-Disposition": {
                type: "string",
                description:
                  "Attachment filename in ASCII `filename=` and UTF-8 `filename*=` forms, for example `Artist_Album_01_Title.mp3`.",
              },
              "Content-Length": {
                type: "integer",
                minimum: 0,
                description: "Complete audio-file size in bytes. The header is omitted when the size is unavailable.",
              },
              "Cache-Control": { type: "string", enum: ["no-store"], description: "Always `no-store`." },
            },
            content: {
              "audio/mpeg": { schema: { type: "string", format: "binary" } },
              "audio/ogg": { schema: { type: "string", format: "binary" } },
              "audio/flac": { schema: { type: "string", format: "binary" } },
            },
          },
          400: publicErrorResponse("`jamendoId` is not numeric."),
          403: publicErrorResponse("The track's Creative Commons terms do not permit downloads."),
          404: publicErrorResponse("No Creative Commons track exists for this `jamendoId`."),
          502: publicErrorResponse("The upstream Jamendo audio file is unavailable or returned no body."),
          429: publicErrorResponse("This client IP exceeded `10` requests in a rolling `60`-second window."),
        },
      },
    },
    async (request, reply) => {
      if (!isInternalRequest(request)) {
        const rateLimit = apiRateLimiter.check(request.ip);
        if (rateLimit.limited) {
          return sendRateLimitError(reply, rateLimit);
        }
      }

      const { jamendoId } = request.params;
      const format = parseJamendoAudioFormat(request.query.format);

      const track = await getCcTrack(jamendoId);
      if (!track?.streamUrl) {
        return reply.status(404).send({ error: "TRACK_NOT_FOUND", message: "No CC track found for this Jamendo id." });
      }
      if (!track.downloadAllowed) {
        return reply.status(403).send({ error: "DOWNLOAD_NOT_ALLOWED", message: "This track is not downloadable." });
      }

      let upstream: Response;
      try {
        upstream = await fetch(swapStreamFormat(track.streamUrl, format));
      } catch (err) {
        log.debug("CcDownload", "Jamendo audio fetch failed:", err instanceof Error ? err.message : String(err));
        return reply.status(502).send({ error: "UPSTREAM_UNAVAILABLE", message: "Could not reach the audio." });
      }
      if (!upstream.ok || !upstream.body) {
        return reply.status(502).send({ error: "UPSTREAM_ERROR", message: "The audio is unavailable." });
      }

      const filename = buildCcDownloadFilename({
        artist: track.artistName,
        album: track.albumName,
        trackNumber: track.albumPosition,
        title: track.title,
        format,
      });

      reply.header("Content-Type", JAMENDO_FORMAT_META[format].mime);
      reply.header("Content-Disposition", contentDisposition(filename));
      const length = upstream.headers.get("content-length");
      if (length) reply.header("Content-Length", length);
      reply.header("Cache-Control", "no-store");

      return reply.send(Readable.fromWeb(upstream.body as NodeWebReadableStream<Uint8Array>));
    },
  );
}