# apps/developer/src/lib/__fixtures__/public-openapi.json
{
  "components": {
    "schemas": {
...
      "CcGenreBrowseResponse": {
        "additionalProperties": false,
        "description": "Creative-Commons genre browse response returned for the exact query `genre:?`.",
        "properties": {
          "genres": {
            "description": "Selectable genres for a follow-up `genre:<name>` CC query.",
            "items": {
              "$ref": "#/components/schemas/CcGenreTile"
            },
            "type": "array"
          },
          "status": {
            "description": "Always `genre-browse` for this response shape.",
            "enum": ["genre-browse"],
            "type": "string"
          }
        },
        "required": ["status", "genres"],
        "type": "object"
      },
      "CcGenreSearchResponse": {
        "additionalProperties": false,
        "description": "Creative-Commons genre results. Select a row by sending its `id` as `selectedCandidate` to `POST /api/v1/cc/resolve`.",
        "properties": {
          "query": {
            "additionalProperties": false,
            "description": "Normalized genre filters used for this search.",
            "properties": {
              "albums": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "The key is always included. Its value is the requested album count, or `null` when albums were not requested."
              },
              "artists": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "The key is always included. Its value is the requested artist count, or `null` when artists were not requested."
              },
              "genres": {
                "description": "Requested normalized genre tags.",
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              "tracks": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "The key is always included. Its value is the requested track count, or `null` when tracks were not requested."
              },
              "vibe": {
                "description": "Normalized ordering mode: `hot` selects the highest-ranked results, while `mixed` samples across the ranked result pool.",
                "enum": ["hot", "mixed"],
                "type": "string"
              }
            },
            "required": ["genres", "vibe", "tracks", "albums", "artists"],
            "type": "object"
          },
          "results": {
            "additionalProperties": false,
            "description": "Candidate lists grouped by requested resource type.",
            "properties": {
              "albums": {
                "anyOf": [
                  {
                    "items": {
                      "$ref": "#/components/schemas/CcGenreAlbumCandidate"
                    },
                    "type": "array"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "The key is always included. Its value is the album-candidate array, or `null` when albums were not requested."
              },
              "artists": {
                "anyOf": [
                  {
                    "items": {
                      "$ref": "#/components/schemas/CcGenreArtistCandidate"
                    },
                    "type": "array"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "The key is always included. Its value is the artist-candidate array, or `null` when artists were not requested."
              },
              "tracks": {
                "anyOf": [
                  {
                    "items": {
                      "$ref": "#/components/schemas/CcGenreTrackCandidate"
                    },
                    "type": "array"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "The key is always included. Its value is the track-candidate array, or `null` when tracks were not requested."
              }
            },
            "required": ["tracks", "albums", "artists"],
            "type": "object"
          },
          "status": {
            "description": "Always `genre-search` for this response shape.",
            "enum": ["genre-search"],
            "type": "string"
          },
          "warnings": {
            "description": "Query-normalization notes; an empty array means no adjustment was needed.",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": ["status", "query", "results", "warnings"],
        "type": "object"
      },
...
      "CcResolveSuccess": {
        "description": "Successful Creative Commons resolve, discriminated by `type` into track, album, or artist payloads.",
        "oneOf": [
          {
            "$ref": "#/components/schemas/CcTrackResolveSuccess"
          },
          {
            "$ref": "#/components/schemas/CcAlbumResolveSuccess"
          },
          {
            "$ref": "#/components/schemas/CcArtistResolveSuccess"
          }
        ]
      },
...
      "CcTrackResolveSuccess": {
        "additionalProperties": false,
        "description": "Successful Creative Commons track resolve.",
        "properties": {
          "id": {
            "description": "Persisted musiccloud CC-track ID for correlation. No public endpoint accepts it; use the final path segment of `shortUrl` with `GET /api/v1/share/{shortId}`.",
            "type": "string"
          },
          "shortUrl": {
            "description": "Canonical musiccloud share URL for this Creative Commons track.",
            "format": "uri",
            "type": "string"
          },
          "track": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CcTrack"
              },
              {
                "required": ["vinylLayout"],
                "type": "object"
              }
            ],
            "description": "Resolved Jamendo track with its vinyl-layout lookup state."
          },
          "type": {
            "description": "Discriminator: always `cc-track` for this response shape.",
            "enum": ["cc-track"],
            "type": "string"
          }
        },
        "required": ["type", "id", "shortUrl", "track"],
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Standard public error envelope. Route-specific error schemas add documented fields such as form validation issues or readiness details.",
        "example": {
          "context": {
            "limit": 10,
            "retryAfterSeconds": 42,
            "windowSeconds": 60
          },
          "error": "MC-API-0003",
          "errorId": "7d33e012-685f-4f11-94da-c6bc72918d7b",
          "message": "Too many requests. You can make 10 requests per 60 seconds. Please try again in 42 seconds. (MC-API-0003)"
        },
        "properties": {
          "context": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ]
            },
            "description": "Optional structured values associated with the error code. For `MC-API-0003`, the object can contain the active limit, window, and retry delay. The key is omitted when no structured values apply.",
            "properties": {
              "limit": {
                "description": "Maximum requests allowed by the rate-limit rule that rejected this request. The key is omitted for non-rate-limit errors.",
                "type": "number"
              },
              "retryAfterSeconds": {
                "description": "Seconds until the client can retry after a `429 Too Many Requests` response. The key is omitted when no retry delay applies.",
                "type": "number"
              },
              "windowSeconds": {
                "description": "Length in seconds of the rate-limit window that rejected this request. The key is omitted for non-rate-limit errors.",
                "type": "number"
              }
            },
            "type": "object"
          },
          "error": {
            "description": "Stable musiccloud error code for programmatic handling, for example `MC-URL-0003`, `MC-API-0003`, or `MC-RES-0001`.",
            "type": "string"
          },
          "errorId": {
            "description": "Unique incident identifier included in the matching backend log. Send it to musiccloud support when reporting the failure.",
            "format": "uuid",
            "type": "string"
          },
          "message": {
            "description": "Safe English failure detail. The final parenthesized value repeats the `error` code.",
            "type": "string"
          }
        },
        "required": ["error", "message", "errorId"],
        "type": "object"
      },
...
      "ResolveDisambiguation": {
        "additionalProperties": false,
        "description": "Resolve could not pick one match; the caller must choose from `candidates`.",
        "example": {
          "candidates": [
            {
              "albumName": "Hunting High and Low",
              "artists": ["a-ha"],
              "artworkUrl": "https://i.scdn.co/image/ab67616d0000b273e58a0f7f1f2f8e4f6a3c8b2d",
              "id": "spotify:2WfaOiMkCvy7F5fcp2zZ8L",
              "title": "Take on Me"
            },
            {
              "albumName": "MTV Unplugged — Summer Solstice",
              "artists": ["a-ha"],
              "artworkUrl": "https://i.scdn.co/image/ab67616d0000b27309d2e5cb5e5a8f6a3c8b2d0a",
              "id": "spotify:4VqPOruhp5EdPBeR92t6lQ",
              "title": "Take on Me (MTV Unplugged)"
            }
          ],
          "status": "disambiguation"
        },
        "properties": {
          "candidates": {
            "description": "Candidate matches to present to the user for selection.",
            "items": {
              "$ref": "#/components/schemas/DisambiguationCandidate"
            },
            "type": "array"
          },
          "status": {
            "description": "Always `disambiguation`; choose one item from `candidates` before resolving again.",
            "enum": ["disambiguation"],
            "type": "string"
          }
        },
        "required": ["status", "candidates"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "title": "musiccloud API",
    "version": "2.1.3"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/cc/download/{jamendoId}": {
      "get": {
        "description": "Returns a complete Creative Commons track as a downloadable attachment. `Content-Disposition` provides an ASCII fallback and a UTF-8 filename following `Artist_Album_NN_Title.ext`; unavailable album and track-position segments are omitted. `format=mp31` selects MP3 at `96 kbps`, `format=mp32` selects MP3 at approximately `256 kbps`, `format=ogg` selects Ogg Vorbis, and `format=flac` selects lossless FLAC. The default is `mp32`. Check `track.downloadAllowed` before offering this action; the endpoint returns `403` when downloads are not permitted.",
        "parameters": [
          {
            "description": "Jamendo delivery format: `mp31` (MP3 96 kbps), `mp32` (MP3 about 256 kbps), `ogg`, or `flac`. Invalid values also fall back to `mp32`.\n\n**Default**: `mp32`.",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "mp32",
              "type": "string"
            }
          },
          {
            "description": "Numeric Jamendo track ID. Read `track.jamendoId`, `album.tracks[].jamendoId`, or `artist.topTracks[].jamendoId` from a successful `POST /api/v1/cc/resolve` or `GET /api/v1/share/{shortId}` response.",
            "in": "path",
            "name": "jamendoId",
            "required": true,
            "schema": {
              "maxLength": 32,
              "minLength": 1,
              "pattern": "^[0-9]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "audio/flac": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "audio/mpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "audio/ogg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Complete Creative Commons audio file served as an attachment.",
            "headers": {
              "Cache-Control": {
                "description": "Always `no-store`.",
                "schema": {
                  "enum": ["no-store"],
                  "type": "string"
                }
              },
              "Content-Disposition": {
                "description": "Attachment filename in ASCII `filename=` and UTF-8 `filename*=` forms, for example `Artist_Album_01_Title.mp3`.",
                "schema": {
                  "type": "string"
                }
              },
              "Content-Length": {
                "description": "Complete audio-file size in bytes. The header is omitted when the size is unavailable.",
                "schema": {
                  "minimum": 0,
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`jamendoId` is not numeric."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The track's Creative Commons terms do not permit downloads."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No Creative Commons track exists for this `jamendoId`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "This client IP exceeded `10` requests in a rolling `60`-second window.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error. Use `errorId` from the response when reporting the failure."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The upstream Jamendo audio file is unavailable or returned no body."
          }
        },
        "summary": "Download a CC track as a correctly named audio file",
        "tags": ["CC"]
      }
    },
...
    "/api/v1/cc/resolve": {
      "post": {
        "description": "Use `query` for Creative-Commons free text, a `title:`/`artist:`/`album:` structured search, or `genre:` discovery. Pasted streaming-service URLs are not accepted. A text search returns `ResolveDisambiguation`; choose one row and send its opaque `candidates[].id` back as `selectedCandidate`. A CC genre result already returns candidate IDs in `results.*[].id`; pass one unchanged as `selectedCandidate` to receive a persisted `cc-track`, `cc-album`, or `cc-artist` response.",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "example1": {
                  "value": {
                    "query": "ambient piano"
                  }
                },
                "example2": {
                  "value": {
                    "selectedCandidate": "jamendo:123456"
                  }
                }
              },
              "schema": {
                "additionalProperties": false,
                "description": "Exactly one of `query` or `selectedCandidate` must be present. Supplying both is rejected with `400`.",
                "oneOf": [
                  {
                    "required": ["query"]
                  },
                  {
                    "required": ["selectedCandidate"]
                  }
                ],
                "properties": {
                  "query": {
                    "description": "Creative-Commons free text, structured search beginning with `title:`, `artist:`, or `album:`, or a `genre:` discovery query. Streaming-service URLs are not accepted.",
                    "example": "ambient piano",
                    "maxLength": 500,
                    "minLength": 1,
                    "type": "string"
                  },
                  "selectedCandidate": {
                    "description": "Opaque candidate token returned as `candidates[].id` or `results.*[].id` by this endpoint. Pass it unchanged; prefixes such as `jamendo:`, `jamendo-album:`, and `jamendo-artist:` identify the response variant.",
                    "example": "jamendo:123456",
                    "maxLength": 200,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "description": "Exactly one of `query` or `selectedCandidate` must be present. Supplying both is rejected with `400`.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "description": "A `ResolveDisambiguation` selection list, a persisted `cc-track`, `cc-album`, or `cc-artist`, or a genre-discovery result with `status` equal to `genre-browse` or `genre-search`.",
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ResolveDisambiguation"
                    },
                    {
                      "$ref": "#/components/schemas/CcResolveSuccess"
                    },
                    {
                      "$ref": "#/components/schemas/CcGenreSearchResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CcGenreBrowseResponse"
                    }
                  ]
                }
              }
            },
            "description": "A `ResolveDisambiguation` selection list, a persisted `cc-track`, `cc-album`, or `cc-artist`, or a genre-discovery result with `status` equal to `genre-browse` or `genre-search`."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The body does not contain exactly one usable `query` or `selectedCandidate`, the query syntax is invalid, or the candidate token has an unsupported format."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing, invalid, or revoked API key."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The selected candidate could not be resolved."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The issued API key exceeded its assigned rolling `60`-second or rolling `24`-hour quota. Inspect `context` and `Retry-After` before retrying.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error. Use `errorId` from the response when reporting the failure."
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Resolve a Creative-Commons free-text or structured query (Jamendo)",
        "tags": ["Resolve"]
      }
    },
...
    "/api/v1/share/{shortId}": {
      "get": {
        "description": "Returns one of six exact share variants selected by `type`: commercial `track`, `album`, or `artist`, and Creative Commons `cc-track`, `cc-album`, or `cc-artist`. Commercial variants contain the matching entity plus `links`. `cc-track` contains only `track`, `og`, and `shortUrl`; `cc-album` and `cc-artist` additionally contain `artistInfo`. Creative Commons variants never contain `links`. Read `type` before accessing the variant-specific entity key.",
        "parameters": [
          {
            "description": "Public musiccloud share code: take the last path segment of `shortUrl` from a successful `POST /api/v1/resolve`, `GET /api/v1/resolve`, or `POST /api/v1/cc/resolve` response.",
            "in": "path",
            "name": "shortId",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharePage"
                }
              }
            },
            "description": "`SharePage`, discriminated by `type` into the exact commercial or Creative-Commons resource shape.",
            "headers": {
              "Cache-Control": {
                "description": "`private, max-age=3600` for commercial `track`, `album`, and `artist` plus `cc-artist`; `no-store` for `cc-track` and `cc-album`.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`shortId` failed validation (empty, too long, or contains disallowed characters)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No commercial or Creative-Commons track, album, or artist exists for this share code."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "This client IP exceeded `10` requests in a rolling `60`-second window.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying the request.",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server error. Use `errorId` from the response when reporting the failure."
          }
        },
        "summary": "Fetch a previously-resolved share",
        "tags": ["Share"],
        "x-codeSamples": [
          {
            "label": "cURL",
            "lang": "Shell",
            "source": "curl \"https://api.musiccloud.io/api/v1/share/aBc123x\""
          },
          {
            "label": "Bash",
            "lang": "Shell",
            "source": "#!/usr/bin/env bash\nset -euo pipefail\n\nBASE_URL=\"https://api.musiccloud.io\"\n\nresponse=$(curl -sS \"$BASE_URL/api/v1/share/aBc123x\")\n\necho \"$response\" | jq ."
          },
          {
            "lang": "JavaScript",
            "source": "const response = await fetch(\"https://api.musiccloud.io/api/v1/share/aBc123x\");\nconst data = await response.json();"
          },
          {
            "lang": "Python",
            "source": "import requests\n\nresponse = requests.get(\"https://api.musiccloud.io/api/v1/share/aBc123x\")\ndata = response.json()"
          },
          {
            "lang": "PHP",
            "source": "<?php\n\n$ch = curl_init(\"https://api.musiccloud.io/api/v1/share/aBc123x\");\ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n\n$response = curl_exec($ch);\ncurl_close($ch);\n$data = json_decode($response, true);"
          },
          {
            "lang": "Ruby",
            "source": "require \"net/http\"\nrequire \"json\"\n\nuri = URI(\"https://api.musiccloud.io/api/v1/share/aBc123x\")\nhttp = Net::HTTP.new(uri.host, uri.port)\nhttp.use_ssl = uri.scheme == \"https\"\n\nrequest = Net::HTTP::Get.new(uri.request_uri)\n\nresponse = http.request(request)\ndata = JSON.parse(response.body)"
          },
          {
            "lang": "Swift",
            "source": "import Foundation\n\nlet url = URL(string: \"https://api.musiccloud.io/api/v1/share/aBc123x\")!\nvar request = URLRequest(url: url)\nrequest.httpMethod = \"GET\"\n\nlet (data, _) = try await URLSession.shared.data(for: request)\nlet json = try JSONSerialization.jsonObject(with: data)"
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.musiccloud.io"
    }
  ],
...
}